Separating the interfaces
